packages/agentspec-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentspec"
version = "0.1.0"
description = "A declarative specification language for defining, documenting, and enforcing AI agent behavior."
requires-python = ">=3.9"
license = "MIT"
keywords = ["ai", "agents", "specification", "enforcement", "safety"]
authors = [
    { name = "AgentSpec Contributors" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
crewai = ["crewai>=0.30.0"]
langchain = ["langchain>=0.1.0", "langchain-core>=0.1.0"]
openai = ["openai>=1.0.0"]
all = ["crewai>=0.30.0", "langchain>=0.1.0", "langchain-core>=0.1.0", "openai>=1.0.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/agentspec/agentspec"
Documentation = "https://agentspec.org"
Repository = "https://github.com/agentspec/agentspec"
Issues = "https://github.com/agentspec/agentspec/issues"

[tool.hatch.build.targets.wheel]
packages = ["agentspec"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.mypy]
python_version = "3.9"
strict = true
plugins = ["pydantic.mypy"]
